# packages/shared/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
...
  /api/v1/tickets:
    get:
      tags:
      - ticket-controller
      operationId: list
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: type
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMapStringObject"
    post:
      tags:
      - ticket-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicket"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/tickets/{id}/resolve:
    post:
      tags:
      - ticket-controller
      operationId: resolve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/tickets/{id}/escalate:
    post:
      tags:
      - ticket-controller
      operationId: escalate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/tickets/{id}/assign:
    post:
      tags:
      - ticket-controller
      operationId: assign
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTicket"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/risk-events/{id}/resolve:
    post:
      tags:
      - risk-controller
      operationId: resolve_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/member-tasks/{id}/complete:
    post:
      tags:
      - member-task-controller
      operationId: complete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/invite-codes/rotate:
    post:
      tags:
      - invite-code-controller
      operationId: rotate
      parameters:
      - name: memberNo
        in: query
        required: false
        schema:
          type: string
      - name: projectScope
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/followups:
    get:
      tags:
      - follow-up-controller
      operationId: list_8
      parameters:
      - name: queue
        in: query
        required: false
        schema:
          type: string
      - name: projectId
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMapStringObject"
    post:
      tags:
      - follow-up-controller
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUp"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/followups/{id}/complete:
    post:
      tags:
      - follow-up-controller
      operationId: complete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteFollowUp"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/invite-codes/mine:
    get:
      tags:
      - invite-code-controller
      operationId: mine
      parameters:
      - name: memberNo
        in: query
        required: false
        schema:
          type: string
      - name: projectScope
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
...
    ApiResponseMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
...
    CreateTicket:
      required:
      - ticketType
      type: object
      properties:
        ticketType:
          type: string
        memberNo:
          type: string
        projectId:
          type: string
        city:
          type: string
        priority:
          type: string
        description:
          type: string
    AssignTicket:
      type: object
      properties:
        assigneeEmployeeId:
          type: integer
          format: int64
...
    ResolveRequest:
      type: object
      properties:
        resolution:
          type: string
...
    CreateFollowUp:
      required:
      - category
      - memberNo
      type: object
      properties:
        memberNo:
          type: string
        projectId:
          type: string
        category:
          type: string
        priority:
          type: string
        content:
          type: string
        assignee:
          type: string
        remindAt:
          type: string
        writeBack:
          type: boolean
        queue:
          type: string
    CompleteFollowUp:
      type: object
      properties:
        resultNote:
          type: string
...
    ApiResponseListMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
...